moved to sqlviewmanager
